// apps/desktop/src/ipc/generated/ScheduleDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ScheduleSourceInfo } from "./ScheduleSourceInfo";

export type ScheduleDto = { id: string, name: string, status: string, strictValidation: boolean, dayDurationMinutes: number, source: ScheduleSourceInfo | null, rev: number, };